Improvements and debug

+++ app/front/controllers/ProfileController.class.php
@@ -84,9 +84,9 @@
           }
 
           if (!isset($_SESSION['errors'])) {
-              header('Location: '.BASE_URL.'profile');
+              header('Location: '.BASE_URL.'profile/show/'.$_SESSION['user']['username']);
           } else {
-              header('Location: '.BASE_URL.'profile/edit');
+              header('Location: '.BASE_URL.'profile/edit'.$_SESSION['user']['username']);
           }
       }
 
@@ -113,9 +113,9 @@
           }
 
           if (!isset($_SESSION['errors'])) {
-              header('Location: '.BASE_URL.'profile');
+              header('Location: '.BASE_URL.'profile/show/'.$_SESSION['user']['username']);
           } else {
-              header('Location: '.BASE_URL.'profile/edit');
+              header('Location: '.BASE_URL.'profile/edit'.$_SESSION['user']['username']);
           }
       }
 
@@ -142,9 +142,9 @@
           }
 
           if (!isset($_SESSION['errors'])) {
-              header('Location: '.BASE_URL.'profile');
+              header('Location: '.BASE_URL.'profile/show/'.$_SESSION['user']['username']);
           } else {
-              header('Location: '.BASE_URL.'profile/edit');
+              header('Location: '.BASE_URL.'profile/edit'.$_SESSION['user']['username']);
           }
       }
 
@@ -171,9 +171,9 @@
           }
 
           if (!isset($_SESSION['errors'])) {
-              header('Location: '.BASE_URL.'profile');
+              header('Location: '.BASE_URL.'profile/show/'.$_SESSION['user']['username']);
           } else {
-              header('Location: '.BASE_URL.'profile/edit');
+              header('Location: '.BASE_URL.'profile/edit'.$_SESSION['user']['username']);
           }
       }
 
@@ -208,9 +208,9 @@
           }
 
           if (!isset($_SESSION['errors'])) {
-              header('Location: '.BASE_URL.'profile');
+              header('Location: '.BASE_URL.'profile/show/'.$_SESSION['user']['username']);
           } else {
-              header('Location: '.BASE_URL.'profile/edit');
+              header('Location: '.BASE_URL.'profile/edit'.$_SESSION['user']['username']);
           }
       }
 
@@ -224,12 +224,15 @@
            $user = new Users();
            $user = $user->populate(['id' => $_SESSION['user']['id']]);
 
+
            try {
-               unlink(UPLOADS_DIR_USERS.$user->getUserImg());
+               if ($user->getUserImg() !== BASE_AVATAR) {
+                   unlink(UPLOADS_DIR_USERS.$user->getUserImg());
+               }
                $user->setUserImg($_FILES['user_img']);
-           } catch (\Exception $e) {
-             array_push($_SESSION['errors'], $e->getMessage());
-           }
+               } catch (\Exception $e) {
+                   array_push($_SESSION['errors'], $e->getMessage());
+               }
 
            try {
              $user->save();
@@ -238,9 +241,9 @@
            }
 
            if (!isset($_SESSION['errors'])) {
-               header('Location: '.BASE_URL.'profile');
+               header('Location: '.BASE_URL.'profile/show/'.$_SESSION['user']['username']);
            } else {
-               header('Location: '.BASE_URL.'profile/edit');
+               header('Location: '.BASE_URL.'profile/edit'.$_SESSION['user']['username']);
            }
        }
